requirements.txt
wheel==0.45.1
#cmake==3.31.1
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.2
typing_extensions==4.12.2
click==8.1.7
dlib==19.24.6
face_recognition==1.3.0
face-recognition-models==0.3.0
numpy==2.1.3
pillow==11.0.0
